pyproject.toml
[tool.poetry]
name = "sp_ask_running_total_daily_report"
version = "0.1.0"
description = ""
authors = ["Guinsly Mondesir <guinslym@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pdbpp = "^0.10.2"
lh3api = "^0.2.0"
ask_academic_dates = "^0.2.3"
pandas = "^0.25.3"
ask_schools = "^0.2.4"
xlrd = "^1.2"
XlsxWriter = "^1.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pdbpp = "^0.10.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
